Allow use of a custom schema director
This is very useful when using hookup in conjunction with a multiple schema
migration solution. The standalone_migration gem is introducing this
functionality in the near future and it'd be good to have support for this use
case ready to go.
________

Rewritten to be more inline with the spirt of existing code, also added
a directory check to avoid weirdness, and named parameter support.